Separate and improve stats counter
Move FPS stats counter out of the video renderer to a separate class.

Count frames when surface is released after rendering, rather than after
ExoPlayer has processed the buffer. This should be slightly more
accurate.
