+++ app/src/main/java/dev/hadrosaur/videodecodeencodedemo/VideoHelpers/VideoMediaCodecVideoRenderer.kt
@@ -49,30 +49,8 @@ class VideoMediaCodecVideoRenderer(
         -1
     )  {
 
-    // Stats counters
-    private var fpsDecodeCounter = 0
-    private var fpsTotalDecodeCounter = 0
-    private var fpsLastMeasuredTime = 0L
-    private var fpsLastLoggedTime = 0L
     private var droppedFrames = 0
     private var lastPresentTime = 0L
-
-    // Set up an int array for fps stats to get an idea of choppiness. Ex. 0-55fps+, 12 buckets
-    // 0-4, 5-9, 10-14, 15-19, 20-24 . . . 50-54, 55+
-    private val MAX_FPS_STATS = 60 // max fps we care about for stats
-    private val NUM_FPS_BUCKETS = 6
-    private val FPS_BUCKET_SIZE = MAX_FPS_STATS / NUM_FPS_BUCKETS // 10fps
-    private val LAST_FPS_BUCKET_START = MAX_FPS_STATS - FPS_BUCKET_SIZE // 50+
-    private val fpsBuckets = IntArray(NUM_FPS_BUCKETS) { 0 }
-
-    // Keep track of mix/max fps
-    private var minFps = MAX_FPS_STATS
-    private var maxFps = 0
-
-    // Choppiness = num frames < 30fps
-    private val CHOPPINESS_CUTOFF = 30
-    private val TOO_MANY_CHOPPY_FRAMES = 10
-    private var numChoppyFrames = 0
 
     /**
      * Keep track of dropped frames
@@ -178,86 +156,19 @@ class VideoMediaCodecVideoRenderer(
         return processSuccess
     }
 
-    private fun resetStatsCounters() {
-        minFps = MAX_FPS_STATS
-        maxFps = 0
-        numChoppyFrames = 0
-        for (i in 0 until NUM_FPS_BUCKETS) {
-            fpsBuckets[i] = 0
-        }
-    }
 
     /**
-     * Adds some logging after each buffer processed to keep track of decode
+     * Update media clock after each buffer processed
      */
     override fun onProcessedOutputBuffer(presentationTimeUs: Long) {
-        val currentTime = System.currentTimeMillis()
-        // If this is the first frame, don't calculate stats
-        if (fpsLastLoggedTime == 0L || fpsLastMeasuredTime == 0L) {
-            fpsLastLoggedTime = currentTime
-            fpsLastMeasuredTime = currentTime
-        } else {
-            fpsDecodeCounter++
-            fpsTotalDecodeCounter++
-
-            // viewModel.updateLog("I have decoded ${decodeCounter} video frames.")
-            if (fpsLastMeasuredTime == currentTime) {
-                fpsLastMeasuredTime -= 3 // 0ms since last frame, subtract time to avoid divide by 0
-            }
-            val currentFrameFps = 1000 / (currentTime - fpsLastMeasuredTime)
-
-
-            // Calculate stats for this frame
-            minFps = getMin(minFps, currentFrameFps.toInt())
-            maxFps = getMax(maxFps, currentFrameFps.toInt())
-
-            if (currentFrameFps < CHOPPINESS_CUTOFF) {
-                numChoppyFrames++
-            }
-
-            // Place this frame's fps in the bucket
-            val currentfpsBucketIndex = getMin(currentFrameFps.toInt(), MAX_FPS_STATS - 1) / FPS_BUCKET_SIZE
-            fpsBuckets[currentfpsBucketIndex]++
-
-
-            // If this is a logging frame, output states and reset counters
-            if (fpsDecodeCounter % LOG_VIDEO_EVERY_N_FRAMES == 0) {
-                val averageFps = fpsDecodeCounter / ((currentTime - fpsLastLoggedTime) / 1000.0)
-
-                val averageFpsString = String.format("%.2f", averageFps)
-                val choppyString = if (numChoppyFrames >= TOO_MANY_CHOPPY_FRAMES) " ---CHOPPY---" else ""
-
-                // FPS buckets line
-                var bucketsString1 = ""
-                for (i in 0 until NUM_FPS_BUCKETS) {
-                    if (i == NUM_FPS_BUCKETS -1) {
-                        // Last bucket
-                        bucketsString1 += "[${i * FPS_BUCKET_SIZE}+: ${fpsBuckets[i]}]"
-                    } else {
-                        bucketsString1 += "[${i * FPS_BUCKET_SIZE}-${(i+1) * FPS_BUCKET_SIZE - 1}: ${fpsBuckets[i]}]    "
-                    }
-                }
-                if (streamNumber == 0)
-                    viewModel.updateLog("\n")
-
-                val logString = "V${streamNumber + 1}@frame $fpsTotalDecodeCounter. FPS: min: ${minFps} max: ${maxFps} avg: ${averageFpsString}. Choppy frames: ${numChoppyFrames}${choppyString}.\n\t" +
-                        bucketsString1
-                viewModel.updateLog(logString)
-
-                fpsLastLoggedTime = currentTime // Update for next fps measurement
-                fpsDecodeCounter = 0
-                resetStatsCounters()
-            }
-        }
-
+        // Check if decoder is stuck
+        // TODO: is this needed for anything?
         if (lastPresentTime == presentationTimeUs && lastPresentTime != 0L) {
             viewModel.updateLog("Last present time is current present time. Frame is stuck! Time: ${presentationTimeUs}")
         }
-
         lastPresentTime = presentationTimeUs
-        mediaClock.updateLastProcessedFrame(presentationTimeUs) // Update media clock with last frame
 
-        fpsLastMeasuredTime = currentTime // Update for next fps measurement
+        mediaClock.updateLastProcessedFrame(presentationTimeUs)
         super.onProcessedOutputBuffer(presentationTimeUs)
     }
